reorg mag_proxy into nice layout
