+++ magic_bridge/ic/src/magic_bridge/src/api/admin.rs
@@ -1,0 +1,7 @@
+use crate::magic::STATE;
+
+pub fn is_authorized() -> Result<(), String> {
+    STATE.with(|s| s.is_authorized())
+}
+
+
