+++ magic_bridge/ic/src/magic_bridge/src/api/init.rs
@@ -1,0 +1,8 @@
+use ic_kit::{ic, macros::init};
+
+use crate::magic::STATE;
+
+#[init]
+fn init() {
+    STATE.with(|s| s.controllers.borrow_mut().push(ic::caller()));
+}
