+++ magic_bridge/ic/src/magic_bridge/src/api/inspect_message.rs
@@ -1,7 +1,7 @@
 use ic_cdk::api::call::accept_message;
 use ic_kit::macros::inspect_message;
 
-use crate::magic::is_authorized;
+use crate::api::admin::is_authorized;
 
 #[inspect_message]
 fn inspect_message() {
