+++ magic_bridge/ic/src/magic_bridge/src/api/mod.rs
@@ -1,0 +1,5 @@
+mod admin;
+mod create;
+mod init;
+mod inspect_message;
+mod upgrade;
