+++ magic_bridge/ic/src/magic_bridge/src/api/upgrade.rs
