+++ magic_bridge/ic/src/magic_bridge/src/magic.rs
@@ -1,12 +1,11 @@
-use crate::factory::{CreateCanisterParam, Factory};
 use crate::types::*;
-use ic_kit::candid::{candid_method, CandidType, Deserialize, Nat};
+use ic_kit::candid::{CandidType, Deserialize};
 use ic_kit::interfaces::management::{
     CanisterStatus, CanisterStatusResponse, DeleteCanister, DepositCycles, InstallCode,
     StartCanister, StopCanister, UninstallCode, UpdateSettings, WithCanisterId,
 };
 use ic_kit::interfaces::Method;
-use ic_kit::{ic, interfaces::management, macros::*};
+use ic_kit::{ic, interfaces::management};
 use ic_kit::{Principal, RejectionCode};
 use management::{InstallCodeArgument, UpdateSettingsArgument};
 
@@ -20,8 +19,8 @@ thread_local! {
 
 #[derive(CandidType, Deserialize, Default)]
 pub struct MagicState {
-    canisters: RefCell<HashMap<EthereumAddr, CanisterId>>,
-    controllers: RefCell<Vec<Principal>>,
+    pub canisters: RefCell<HashMap<EthereumAddr, CanisterId>>,
+    pub controllers: RefCell<Vec<Principal>>,
 }
 
 #[derive(CandidType, Deserialize, Default)]
@@ -139,63 +138,4 @@ impl MagicState {
         self.canisters.replace(stable_magic_state.canisters);
         self.controllers.replace(stable_magic_state.controllers);
     }
-}
-
-pub fn is_authorized() -> Result<(), String> {
-    STATE.with(|s| s.is_authorized())
-}
-
-#[init]
-fn init() {
-    // todo: add each of the proxies as controllers
-    STATE.with(|s| s.controllers.borrow_mut().push(ic::caller()));
-}
-
-#[update(name = "create", guard = "is_authorized")]
-#[candid_method(update, rename = "create")]
-async fn create(eth_addr: Principal, token_type: TokenType, payload: Vec<Nat>) -> MagicResponse {
-    let self_id = ic::id();
-    let caller = ic::caller();
-    let canister_exits = STATE.with(|s| s.get_canister(eth_addr));
-
-    let canister_id = if let Some(canister_id) = canister_exits {
-        canister_id
-    } else {
-        let logo = String::from("/s");
-        let name = str::from_utf8(&payload[3].0.to_bytes_be()[..])
-            .unwrap()
-            .to_string();
-        let symbol = str::from_utf8(&payload[4].0.to_bytes_be()[..])
-            .unwrap()
-            .to_string();
-        // verify this payload[5] to be a base 10 or not, should be
-        let decimals = u8::from_str_radix(&payload[5].to_string(), 10).unwrap();
-
-        let create_param = CreateCanisterParam {
-            logo,
-            name,
-            symbol,
-            decimals,
-            total_supply: Nat::from(0_u32),
-            owner: caller,
-            controllers: vec![caller, self_id],
-            cycles: 10_000_000_000_000,
-            fee: Nat::from(0_u32),
-            fee_to: ic::id(),
-            cap: Principal::from_text("e22n6-waaaa-aaaah-qcd2q-cai").unwrap(),
-            token_type,
-        };
-
-        let create_canister = Factory::create(create_param).await;
-
-        match create_canister {
-            Ok(canister_id) => {
-                STATE.with(|s| s.insert_canister(eth_addr, canister_id));
-                canister_id
-            }
-            Err(error) => return Err(TxError::Other(format!("FactoryError: {:?}", error))),
-        }
-    };
-
-    Ok(canister_id)
 }
